multi cubes

+++ src/Frustum.h
@@ -8,12 +8,12 @@
 
 #include <glm/glm.hpp>
 
-auto projection_create(int n, int f, int r, int t){
+auto projection_create(float n, float f, float r, float t){
     glm::mat4 projection(0.f);
-    projection[0] = glm::vec4(((float)n)/(float)r, 0,0,0);
-    projection[1] = glm::vec4(0, ((float)n)/(float)t,0,0);
-    projection[2] = glm::vec4(0, 0,((float)(n+f))/(float)(n-f), -1);
-    projection[3] = glm::vec4(0, 0,((float)(2*n*f))/(float)(n-f),0);
+    projection[0] = glm::vec4(n/r, 0,0,0);
+    projection[1] = glm::vec4(0, n/t,0,0);
+    projection[2] = glm::vec4(0, 0,(n+f)/(n-f), -1);
+    projection[3] = glm::vec4(0, 0,(2*n*f)/(n-f),0);
     return projection;
 }
